Replace QStringLiteral by QLatin1String for Qt < 5
In Qt 4 there is no class QStringLiteral. A less performant replacement is QLatin1String. But this class does not have an arg() method. Therefore, some occurences of QStringLiteral had to be replaced by QString. All other occurences of QStringLiteral are replaced by QLatin1String globally using a define. In addition, we need some more defines to tell the moc compiler of Qt 4 about the namespaces.